pyproject.toml
[project]
name = "used-car-marketplace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosmtplib>=5.1.0",
    "alembic>=1.18.4",
    "annotated-doc>=0.0.4",
    "annotated-types>=0.7.0",
    "anyio>=4.13.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "blinker>=1.9.0",
    "cffi>=2.0.0",
    "click>=8.3.2",
    "cryptography>=46.0.7",
    "dnspython>=2.8.0",
    "ecdsa>=0.19.2",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "fastapi-mail>=1.6.2",
    "greenlet>=3.4.0",
    "h11>=0.16.0",
    "httptools>=0.7.1",
    "httpx>=0.28.1",
    "idna>=3.11",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "mako>=1.3.10",
    "markupsafe>=3.0.3",
    "passlib>=1.7.4",
    "pyasn1>=0.6.3",
    "pycparser>=3.0",
    "pydantic>=2.13.0",
    "pydantic-core>=2.46.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-keycloak>=7.1.1",
    "python-multipart>=0.0.26",
    "pyyaml>=6.0.3",
    "regex>=2025.11.3",
    "rsa>=4.9.1",
    "six>=1.17.0",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.38",
    "starlette>=0.52.1",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.2",
    "uvicorn>=0.44.0",
    "uvloop>=0.22.1",
    "watchfiles>=1.1.1",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.10",
]
